+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added: Interface to customize preset description.
 - Added: It's now possible to save rdvgame files for race games. This is not available for multiworld.
 - Added: When editing a Pickup Node, there's now a button to find an unused pickup index.
+- Changed: Discord login is now performed via your browser, instead of the Discord client.
 - Changed: Door Lock mode Two-way is now named Doors. The functionality is unchanged.
 - Changed: Improved preset descriptions, making them significantly simpler.
 - Changed: Some preset options which are not ready for wide consumption have been hidden by default. To show all preset options, please select `Advanced > Show Experimental Settings`.